fix(meta-dev): generate u-boot-env using mkenvimage and deploy
Fix uboot.env not being included on the boot partition.

The boot process is a follows: U-boot looks for a hard-coded
file called "uboot.env" on the first FAT partition. This file
must be created using mkenvimage, which adds a CRC checksum to
the file. This environment defines variables and instructs U-boot
to look for "boot.scr" which is the actual boot script that loads
and boots the kernel.

Signed-off-by: Stefan Gloor <code@stefan-gloor.ch>